requirements.txt
fastapi
uvicorn
google-generativeai
requests
sqlalchemy
python-dotenv
pydantic
